requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Faker==16.3.0
idna==3.4
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.7
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.14
